+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/stickers/CreateNewStaticStickerSet.kt
@@ -2,12 +2,12 @@ package com.github.insanusmokrassar.TelegramBotAPI.requests.stickers
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.requests.common.CommonMultipartFileRequest
-import com.github.insanusmokrassar.TelegramBotAPI.requests.stickers.abstracts.StickerSetAction
+import com.github.insanusmokrassar.TelegramBotAPI.requests.stickers.abstracts.StandardStickerSetAction
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.stickers.MaskPosition
 import kotlinx.serialization.*
 
-fun CreateNewStickerSet(
+fun CreateNewStaticStickerSet(
     userId: UserId,
     name: String,
     sticker: InputFile,
@@ -15,7 +15,7 @@ fun CreateNewStickerSet(
     containsMasks: Boolean? = null,
     maskPosition: MaskPosition? = null
 ): Request<Boolean> {
-    val data = CreateNewStickerSet(userId, name, emojis, sticker as? FileId, containsMasks, maskPosition)
+    val data = CreateNewStaticStickerSet(userId, name, emojis, sticker as? FileId, containsMasks, maskPosition)
     return when (sticker) {
         is MultipartFile -> CommonMultipartFileRequest(
             data,
@@ -25,8 +25,23 @@ fun CreateNewStickerSet(
     }
 }
 
+fun CreateNewStickerSet(
+    userId: UserId,
+    name: String,
+    sticker: InputFile,
+    emojis: String,
+    containsMasks: Boolean? = null,
+    maskPosition: MaskPosition? = null
+): Request<Boolean> = CreateNewStaticStickerSet(userId, name, sticker, emojis, containsMasks, maskPosition)
+
+@Deprecated(
+    "Renamed",
+    ReplaceWith("CreateNewStaticStickerSet", "com.github.insanusmokrassar.TelegramBotAPI.requests.stickers.CreateNewStaticStickerSet")
+)
+typealias CreateNewStickerSet = CreateNewStaticStickerSet
+
 @Serializable
-data class CreateNewStickerSet internal constructor(
+data class CreateNewStaticStickerSet internal constructor(
     @SerialName(userIdField)
     override val userId: UserId,
     @SerialName(nameField)
@@ -39,7 +54,7 @@ data class CreateNewStickerSet internal constructor(
     val containsMasks: Boolean? = null,
     @SerialName(maskPositionField)
     override val maskPosition: MaskPosition? = null
-) : StickerSetAction {
+) : StandardStickerSetAction {
     init {
         if(emojis.isEmpty()) {
             throw IllegalArgumentException("Emojis must not be empty")