+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/stickers/abstracts/StandardStickerSetAction.kt
@@ -1,0 +1,8 @@
+package com.github.insanusmokrassar.TelegramBotAPI.requests.stickers.abstracts
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.stickers.MaskPosition
+
+interface StandardStickerSetAction : StickerSetAction {
+    val emojis: String // must be more than one
+    val maskPosition: MaskPosition?
+}