+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/RawMessage.kt
@@ -19,7 +19,8 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.message.payments.Success
 import com.github.insanusmokrassar.TelegramBotAPI.types.payments.Invoice
 import com.github.insanusmokrassar.TelegramBotAPI.types.payments.SuccessfulPayment
 import com.github.insanusmokrassar.TelegramBotAPI.types.polls.Poll
-import kotlinx.serialization.*
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
 import kotlin.reflect.KClass
 
 // TODO:: add PassportData type
@@ -74,6 +75,7 @@ internal data class RawMessage(
     private val migrate_from_chat_id: ChatIdentifier? = null,
     private val pinned_message: RawMessage? = null,
     private val invoice: Invoice? = null,
+    private val dice: Dice? = null,
     private val successful_payment: SuccessfulPayment? = null,
 
     // login property
@@ -123,6 +125,7 @@ internal data class RawMessage(
                 adaptedCaptionEntities
             )
             sticker != null -> StickerContent(sticker)
+            dice != null -> DiceContent(dice)
             game != null -> GameContent(game.asGame)
             video_note != null -> VideoNoteContent(video_note)
             contact != null -> ContactContent(contact)