+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/DiceContent.kt
@@ -1,0 +1,19 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.message.content
+
+import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
+import com.github.insanusmokrassar.TelegramBotAPI.requests.send.SendDice
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.ContentMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MessageContent
+
+data class DiceContent(
+    val dice: Dice
+) : MessageContent {
+    override fun createResend(
+        chatId: ChatIdentifier,
+        disableNotification: Boolean,
+        replyToMessageId: MessageIdentifier?,
+        replyMarkup: KeyboardMarkup?
+    ): Request<ContentMessage<DiceContent>> = SendDice(chatId, disableNotification, replyToMessageId, replyMarkup)
+}