+++ CHANGELOG.md
@@ -1,5 +1,43 @@
 # TelegramBotAPI changelog
 
+## 0.26.0
+
+* `Common`:
+    * Versions updates:
+        * `Klock`: `1.10.0` -> `1.10.3`
+* `TelegramBotAPI`:
+    * Request `SendDice` was added (calling [sendDice](https://core.telegram.org/bots/api#senddice))
+    * Class `Dice` was added (type [dice](https://core.telegram.org/bots/api#dice))
+    * Class `DiceContent` was added (for including it in [message](https://core.telegram.org/bots/api#message) object)
+    * `BotCommand` was added
+    * `GetMyCommands` request was added
+    * `SetMyCommands` request was added
+    * `GetMe` now is object instead of class
+    * `GetMe` was replaced into package `com.github.insanusmokrassar.TelegramBotAPI.requests.bot.GetMe`
+    * `CreateNewStickerSet` renamed to `CreateStaticNewStickerSet`
+    * `CreateNewAnimatedStickerSet` request was added (it handle work with `tgs_sticker`)
+    * `StickerSet#thumb` was added
+    * `AddStickerToSet` renamed to `AddStaticStickerToSet`
+    * `AddAnimatedStickerToSet` request was added
+    * `SetStickerSetThumb` request was added
+    * Most of sticker actions now implements `StandardStickerSetAction` instead of `StickerSetAction`
+    * `getUpdatesLimit` was added to be ensure in get updates limit
+    * `GetUpdates` now will check count of requesting updates and throw exception if it is not in range `1 .. 100`
+    * `GetUpdates#limit` now is not nullable and by default set up to 100
+* `TelegramBotAPI-extensions-api`:
+    * Extensions `sendDice` was added
+    * Extension `getMyCommands` request was added
+    * Extension `setMyCommands` request was added
+    * Extension `getMe` was replaced into package `com.github.insanusmokrassar.TelegramBotAPI.extensions.api.bot.GetMeKt.getMe`
+    * **All extensions `createNewStickerSet` was renamed to `createNewStaticStickerSet`**
+    * Extensions `createNewAnimatedStickerSet` was added
+    * **All extensions `addStickerToSet` was renamed to `addStaticStickerToSet`**
+    * Extensions `addAnimatedStickerToSet` was added
+    * Extensions `setStickerSetThumb` was added
+    * Extension `startGettingUpdates` now will drop `SentMediaGroupUpdate` in case if it is the last in updates group
+    and size of retrieved updates is equal to 100 (max count of retrieved updates)
+    * Extensions `getUpdates` now will receive only not nullable `limit` parameter
+
 ## 0.25.0
 
 * Common: